+++ docs/Changelog.md
@@ -10,6 +10,7 @@ Nussknacker versions
 
 0.3.0 (not released yet)
 ------------------------
+* [#1261](https://github.com/TouK/nussknacker/pull/1261) Fix: null-safe operator on maps?.foo caused Property or field cannot be found on object
 * [#1244](https://github.com/TouK/nussknacker/pull/1244) Ability to define `variablesToHide` in `Parameter`
 * [#1165](https://github.com/TouK/nussknacker/pull/1165) Typed global variables
 * [#1128](https://github.com/TouK/nussknacker/pull/1128) Union-memo transformer